refactor(sidebar): extract ChangedRow + StatusBadge to own file
Makes ChangedRow reusable from WorkingTreeSectionView (added later in
this branch).
